+++ telemonitor/__main__.py
@@ -1,5 +1,13 @@
-from telemonitor import main
+from telemonitor.core import cli_arguments_parser
+from telemonitor import startup
+
+
+args = cli_arguments_parser()
+
+
+def run():
+    startup.run()
 
 
 if __name__ == "__main__":
-    main.run()
+    run()