+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Telemonitor Changelog
 
 
-## [**3.1.0** Development](https://github.com/maximilionus/Telemonitor/tree/development)
+## [**4.0.0** Development](https://github.com/maximilionus/Telemonitor/tree/dev/4.0.0)
 
 ### Added
 - Feature to start **Telemonitor** with native python virtual environment *(No poetry)*
 - Feature to change path to the `systemd` service launch `.sh` script with `--systemd-service apply` command to merge all changes to service file
+- New command for starting the app with *poetry*: `poetry run telemonitor`. Old `poetry run telem` can still be used
 - Logging system can now be fully disabled with `--no-logging` startup argument
 - Path of the service launch script now will be showed in `--systemd-service status` output
 
@@ -14,6 +15,11 @@
 - Fixed `--systemd-service upgrade` behavior
 
 ### Changed
+- Code structure major changes:
+  - Main launch script is now `__main__.py`
+  - Bot-related part split to `bot.py`
+  - Startup procedure split to `startup.py`
+  - `helpers` module renamed to `core`
 - Systemd `.service` file was upgraded to new version *(See [README](./README.md#systemd-service-control) for upgrade instructions)*
 - Updated project dependencies to newer versions